+++ failfast/src/main/kotlin/failfast/ExecutionListener.kt
@@ -2,6 +2,5 @@ package failfast
 
 interface ExecutionListener {
     suspend fun testStarted(testDescriptor: TestDescription)
-    suspend fun testFinished(testDescriptor: TestDescription, testResult: TestResult)
-
+    suspend fun testFinished(testPlusResult: TestPlusResult)
 }