+++ failfast/src/main/kotlin/failfast/internal/ContextInfo.kt
@@ -2,10 +2,10 @@ package failfast.internal
 
 import failfast.Context
 import failfast.TestDescription
-import failfast.TestResult
+import failfast.TestPlusResult
 import kotlinx.coroutines.Deferred
 
 internal data class ContextInfo(
     val contexts: List<Context>,
-    val tests: Map<TestDescription, Deferred<TestResult>>
+    val tests: Map<TestDescription, Deferred<TestPlusResult>>
 )