+++ failfast/src/main/kotlin/failfast/internal/SingleTestExecutor.kt
@@ -1,6 +1,14 @@
 package failfast.internal
 
-import failfast.*
+import failfast.ContextDSL
+import failfast.ContextLambda
+import failfast.ContextPath
+import failfast.FailFastException
+import failfast.Failed
+import failfast.RootContext
+import failfast.Success
+import failfast.TestLambda
+import failfast.TestResult
 
 /**
  * Executes a single test with all its parent contexts
@@ -16,6 +24,8 @@ internal class SingleTestExecutor(private val context: RootContext, private val 
             throw FailFastException("specified test not found: $test")
         } catch (e: TestResultAvailable) {
             e.testResult
+        } catch (e: Throwable) {
+            Failed(e)
         }
         closeables.forEach { it.close() }
         return testResult
@@ -63,16 +73,12 @@ internal class SingleTestExecutor(private val context: RootContext, private val 
 
         override suspend fun test(name: String, function: TestLambda) {
             if (test.name == name) {
-                val stackTrace =
-                    RuntimeException().stackTrace.first { !(it.fileName?.endsWith("SingleTestExecutor.kt") ?: true) }
-                val testDescription = TestDescription(test, stackTrace)
-
                 throw TestResultAvailable(
                     try {
                         function()
-                        Success(testDescription, (System.nanoTime() - startTime) / 1000)
+                        Success((System.nanoTime() - startTime) / 1000)
                     } catch (e: Throwable) {
-                        Failed(testDescription, e)
+                        Failed(e)
                     }
                 )
             }