+++ failfast/src/test/kotlin/failfast/internal/TestResultFixtures.kt
@@ -4,6 +4,7 @@ import failfast.Context
 import failfast.Failed
 import failfast.Success
 import failfast.TestDescription
+import failfast.TestPlusResult
 
 object TestResultFixtures {
     val rootContext = Context("the test runner", null)
@@ -11,15 +12,20 @@ object TestResultFixtures {
     val subSubContext = Context("deeper", subContext)
     val failure = RuntimeException("failure message\nwith newline")
     private val stackTraceElement = StackTraceElement("ClassName", "method", "file", 123)
-    val testResults = listOf(
-        Success(TestDescription(rootContext, "supports describe/it syntax", stackTraceElement), 10),
-        Success(
-            TestDescription(subContext, "sub-contexts also contain tests", stackTraceElement),
-            20
+    val testResults: List<TestPlusResult> = listOf(
+        TestPlusResult(
+            TestDescription(rootContext, "supports describe/it syntax", stackTraceElement),
+            Success(10)
         ),
-        Failed(
-            TestDescription(subContext, "failed test", stackTraceElement),
-            failure
+        TestPlusResult(
+            TestDescription(subContext, "sub-contexts also contain tests", stackTraceElement), Success(
+                20
+            )
+        ),
+        TestPlusResult(
+            TestDescription(subContext, "failed test", stackTraceElement), Failed(
+                failure
+            )
         )
     )
 }